Modifications et suppressions par le créateur

+++ backend/controllers/sauce.js
@@ -15,41 +15,52 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    const sauceObject = req.file ? 
-        { 
-            ...JSON.parse(req.body.sauce),
-            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    const sauceObject = req.file
+      ? {
+          ...JSON.parse(req.body.sauce),
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
         }
-        : { ...req.body }; 
-    if (req.file) {
-        Sauce.findOne({ _id: req.params.id })
-            .then(sauce => {
-                const filename = sauce.imageUrl.split('/images/')[1];
-                if (fs.existsSync(`images/${filename}`)) {
-                    fs.unlinkSync(`images/${filename}`);
-                }
-            })
-            .catch(mongooseError => res.status(400).json({ mongooseError }));
-    }
+      : { ...req.body };
+  
+    // Vérifier l'autorisation de modification de la sauce
+    Sauce.findOne({ _id: req.params.id })
+      .then(sauce => {
+        // Vérifier si l'utilisateur est autorisé à modifier la sauce
+        if (sauce.userId !== req.userId) {
+          return res.status(401).json({ message: 'Non autorisé à modifier cette sauce.' });
+        }
+  
+        if (req.file) {
+          const filename = sauce.imageUrl.split('/images/')[1];
+          if (fs.existsSync(`images/${filename}`)) {
+            fs.unlinkSync(`images/${filename}`);
+          }
+        }
 
-    Sauce.updateOne({ _id: req.params.id }, {
-        ...sauceObject,
-        _id: req.params.id
-    })
-        .then(() => res.status(200).json({ message: 'Sauce modifiée' }))
-        .catch(mongooseError => res.status(400).json({ mongooseError }));
+        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Sauce modifiée' }))
+          .catch(mongooseError => res.status(400).json({ mongooseError }));
+      })
+      .catch(mongooseError => res.status(400).json({ mongooseError }));
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.findOne({_id: req.params.id})
-    .then(sauce => {
-    const filename = sauce.imageUrl.split('/images/')[1];
-    fs.unlink(`images/${filename}`, () => {
-    Sauce.deleteOne({_id: req.params.id})
-    .then(()=> res.status(200).json({ message: 'Sauce supprimée'}))
-    .catch(mongooseError => res.status(400).json({ mongooseError }))
-    });
-})
+    // Vérifier l'autorisation de suppression de la sauce
+    Sauce.findOne({ _id: req.params.id })
+      .then(sauce => {
+        // Vérifier si l'utilisateur est autorisé à supprimer la sauce
+        if (sauce.userId !== req.userId) {
+          return res.status(401).json({ message: 'Non autorisé à supprimer cette sauce.' });
+        }
+  
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce supprimée' }))
+            .catch(mongooseError => res.status(400).json({ mongooseError }));
+        });
+      })
+      .catch(mongooseError => res.status(400).json({ mongooseError }));
 };
 
 exports.getAllSauces = (req, res, next) => {
